Title: Add calculator class library
Body: This is the initial commit for the calculator class library. The library includes a set of mathematical functions that can be used to develop a simple calculator in Python, WPF, or WinForm. The library is designed to be easy to use and integrate into existing codebases. This is the first step towards creating a valuable tool for developers who need to perform mathematical calculations in their applications.